Fix race condition in buffered blit (caused flicker).
The blit would fire a new job off and _then_ wait for the previous blit to finish. However if the next blit caught up then this could cause two threads to blit concurrently. The wait is now placed before we fire off a new job so only one blit can be in flight. This fixes some intermittent flickering.